Extract parquet logic into submodule from execution::context
